separate io from CLASS functions
