+++ .gitignore
@@ -1,1 +1,4 @@
-class_public-master-new-dmeff/
+/class_public-master-new-dmeff
+/build
+/dist
+/docs/build
