+++ data/planck2018/base_mnu_plikHM_TTTEEE_lowl_lowE_lensing.paramnames
