+++ src/bsavi.egg-info/requires.txt
